refactor(app): extract dashboard container row
